+++ numpy/core/tests/test_regression.py
@@ -14,8 +14,8 @@ from numpy.testing import (
         assert_raises_regex, assert_warns, suppress_warnings,
         _assert_valid_refcount, HAS_REFCOUNT,
         )
+from numpy.testing._private.utils import _no_tracing
 from numpy.compat import asbytes, asunicode, long, pickle
-from test.support import no_tracing
 
 try:
     RecursionError
@@ -1315,7 +1315,7 @@ class TestRegression:
             assert_(pickle.loads(
                 pickle.dumps(test_record, protocol=proto)) == test_record)
 
-    @no_tracing
+    @_no_tracing
     def test_blasdot_uninitialized_memory(self):
         # Ticket #950
         for m in [0, 1, 2]:
